+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `Pubkey.find_program_address()` function, to find a PDA given a list of seeds and the program key
+
 ## [0.2.4]
 
 ### Added